Finish implementing keyboard inputs
I made a system for getting keyboard inputs from the player. All entities that need keyboard input need the "GameControls" component and they can get it. Note when spawning the entity, you need to setup the controls.

TODO: Write a function in the main systems.rs to load keybindings from a .json file